+++ NEWS.md
@@ -11,6 +11,7 @@
 - Ribbons for 90% intervals in plot for FLQuant and FQuants now have a dotted line added.
 - plot(FLSR) has been improved: axis labels, use of plotmath and regression line on residuals AR plot.
 - plot(FLQuant) now allows specification of quantiles
+- plot methods now use year in the xaxis if only one season, keeps using date, as ISOdate, otherwise
 
 ## BUG FIXES
 - Call to dcast had wrong argument name for value.var